Divide init and socket CB.
